+++ test/sub_graph_test.py
@@ -23,9 +23,9 @@
 
 import pytest
 
-from freexgraph import GraphNode, FreExNode, FreExGraph
+from freexgraph import FreExGraph
 
-# Graph used in thoses tests are from valid_complex_graph fixture
+# Graph used in theses tests are from valid_complex_graph fixture
 #
 #
 #            A                      B
@@ -51,30 +51,108 @@ def test_sub_graph_without_to_node(valid_complex_graph, visitor_test):
 
     sub_graph: FreExGraph = valid_complex_graph.sub_graph(from_node_id="F")
 
-    visitor_test.visit(valid_complex_graph)
     visitor_test.visit(sub_graph.root)
+
     assert len(visitor_test.visited) == 7
+    sorted_visit = sorted(visitor_test.visited)
+
+    assert sorted_visit[0] == "F"
+    assert sorted_visit[1] == "H"
+    assert sorted_visit[2] == "I"
+    assert sorted_visit[3] == "J"
+    assert sorted_visit[4] == "K"
+    assert sorted_visit[5] == "L"
+    assert sorted_visit[6] == "M"
 
 
-def test_sub_graph_of_one_element(valid_complex_graph):
-    ...
+def test_sub_graph_of_one_element(valid_complex_graph, visitor_test):
+    sub_graph: FreExGraph = valid_complex_graph.sub_graph(
+        from_node_id="E", to_nodes_id=["E"]
+    )
+
+    visitor_test.visit(sub_graph.root)
+
+    assert len(visitor_test.visited) == 1
+    assert visitor_test.visited[0] == "E"
 
 
-def test_sub_graph_with_to_node_not_found(valid_complex_graph):
-    ...
+def test_sub_graph_with_to_node(valid_complex_graph, visitor_test):
+    # From A to K and J (L and M are not taken as K and J are stopping the subgraph)
+    #
+    #            A
+    #         /     \
+    #        C       D
+    #                 \
+    #                  F .______,
+    #               /  |  \     \
+    #             H    I   J     `K
+    #
+    sub_graph: FreExGraph = valid_complex_graph.sub_graph(
+        from_node_id="A", to_nodes_id=["K", "J"]
+    )
+
+    visitor_test.visit(sub_graph.root)
+
+    assert len(visitor_test.visited) == 8
+    sorted_visit = sorted(visitor_test.visited)
+
+    assert sorted_visit[0] == "A"
+    assert sorted_visit[1] == "C"
+    assert sorted_visit[2] == "D"
+    assert sorted_visit[3] == "F"
+    assert sorted_visit[4] == "H"
+    assert sorted_visit[5] == "I"
+    assert sorted_visit[6] == "J"
+    assert sorted_visit[7] == "K"
 
 
-def test_sub_graph_with_to_node(valid_complex_graph):
-    ...
+def test_sub_graph_with_to_node_not_found(valid_complex_graph, visitor_test):
+    # From F to A, A is not in the sub graph starting from F, so it's like no to_node where set
+    #
+    #                  F .______,
+    #               /  |  \     \
+    #             H    I   J     `,K.
+    #                     /,_____/   \
+    #                    L             M
 
+    sub_graph: FreExGraph = valid_complex_graph.sub_graph(from_node_id="F", to_nodes_id="A")
 
-def test_sub_graph_with_to_node_partial(valid_complex_graph):
-    ...
+    visitor_test.visit(sub_graph.root)
+
+    assert len(visitor_test.visited) == 7
+    sorted_visit = sorted(visitor_test.visited)
+
+    assert sorted_visit[0] == "F"
+    assert sorted_visit[1] == "H"
+    assert sorted_visit[2] == "I"
+    assert sorted_visit[3] == "J"
+    assert sorted_visit[4] == "K"
+    assert sorted_visit[5] == "L"
+    assert sorted_visit[6] == "M"
 
 
 def test_sub_graph_error_on_from_node(valid_complex_graph):
-    ...
+    with pytest.raises(AssertionError) as e:
+        valid_complex_graph.sub_graph(from_node_id="NOT_EXISTING")
+        assert (
+            "Error sub graph from node NOT_EXISTING, node has to be in the execution graph"
+            == e
+        )
 
 
 def test_sub_graph_error_on_to_node(valid_complex_graph):
-    ...
+    with pytest.raises(AssertionError) as e:
+        valid_complex_graph.sub_graph(from_node_id="F", to_nodes_id=["NOT_EXIST"])
+        assert (
+            "Error sub graph to node NOT_EXIST, node has to be in the execution graph"
+            == e
+        )
+
+    with pytest.raises(AssertionError) as e:
+        valid_complex_graph.sub_graph(
+            from_node_id="F", to_nodes_id=["H", "M", "NOT_EXISTING"]
+        )
+        assert (
+            "Error sub graph to node NOT_EXISTING, node has to be in the execution graph"
+            == e
+        )